+++ src/capitulo_4/4.8/cleanup.c
